fix(security): address code review findings for phantom_gh_exec
Review fixes:

1. Block git --config (space-separated form)
   - Added `arg === "--config"` to catch `git --config alias.x=!env x`
   - Previous check only caught `--config=` form

2. Split in-process-tools.ts to meet 300-line limit
   - Extracted GitHub exec logic to src/agent/gh-exec.ts (255 lines)
   - in-process-tools.ts now 128 lines
   - Re-exports from gh-exec.ts for backwards compatibility

Added test for --config space-separated bypass vector.
